cleanup params/version.go
