+++ contracts/testutil/blockchain_parameters.go
@@ -4,13 +4,13 @@ import (
 	"math/big"
 
 	"github.com/celo-org/celo-blockchain/contracts/abis"
-	"github.com/celo-org/celo-blockchain/params"
+	"github.com/celo-org/celo-blockchain/contracts/internal/config"
 )
 
 type BlockchainParametersMock struct {
 	ContractMock
 
-	MinimumVersion                             params.VersionInfo
+	MinimumVersion                             config.VersionInfo
 	BlockGasLimitValue                         *big.Int
 	LookbackWindow                             *big.Int
 	IntrinsicGasForAlternativeFeeCurrencyValue *big.Int
@@ -18,7 +18,7 @@ type BlockchainParametersMock struct {
 
 func NewBlockchainParametersMock() *BlockchainParametersMock {
 	mock := &BlockchainParametersMock{
-		MinimumVersion:     params.VersionInfo{Major: 1, Minor: 0, Patch: 0},
+		MinimumVersion:     config.VersionInfo{Major: 1, Minor: 0, Patch: 0},
 		BlockGasLimitValue: big.NewInt(20000000),
 		LookbackWindow:     big.NewInt(3),
 		IntrinsicGasForAlternativeFeeCurrencyValue: big.NewInt(10000),
