+++ params/version.go
@@ -31,45 +31,9 @@ const (
 	VersionMeta  = "unstable" // Version metadata to append to the version string
 )
 
-type VersionInfo struct {
-	Major uint64
-	Minor uint64
-	Patch uint64
-}
-
-// Cmp compares x and y and returns:
-//
-//   -1 if x <  y
-//    0 if x == y
-//   +1 if x >  y
-//
-func cmp(x uint64, y uint64) int {
-	if x < y {
-		return -1
-	}
-	if x > y {
-		return 1
-	}
-	return 0
-}
-
-func (v *VersionInfo) Cmp(version *VersionInfo) int {
-	if v.Major == version.Major {
-		if v.Minor == version.Minor {
-			return cmp(v.Patch, version.Patch)
-		}
-		return cmp(v.Minor, version.Minor)
-	}
-	return cmp(v.Major, version.Major)
-}
-
 // Version holds the textual version string.
 var Version = func() string {
 	return fmt.Sprintf("%d.%d.%d", VersionMajor, VersionMinor, VersionPatch)
-}()
-
-var CurrentVersionInfo = func() *VersionInfo {
-	return &VersionInfo{VersionMajor, VersionMinor, VersionPatch}
 }()
 
 // VersionWithMeta holds the textual version string including the metadata.
